% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whipala.R
\name{whipala}
\alias{whipala}
\title{Generador de whipalas de diferentes dimensiones}
\usage{
whipala(ncol = 7, nrow, plot = TRUE)
}
\arguments{
\item{ncol}{Número de columnas de la whipala deseada.}

\item{nrow}{Número de líneas de la whipala deseada. La whipala es cuadrada entonces este valor no es necesario a priori. Si sólo se pide ncol nrow será el mismo valor}

\item{plot}{Lógico. Para ver el gráfico}
}
\description{
Generador de whipalas de diferentes dimensiones
}
\examples{
whipala()
whipala(12)
}
