% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whipalette.R
\docType{data}
\name{whipalette}
\alias{whipalette}
\title{Los colores de la whipala}
\format{
An object of class \code{character} of length 7.
}
\usage{
whipalette
}
\description{
Colores baseados en figura de Huhsunqu, Public Domain, https://commons.wikimedia.org/w/index.php?curid=988731
}
\keyword{datasets}
