% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whipala.R
\docType{data}
\name{whipalette}
\alias{whipalette}
\title{Los colores de la whipala}
\format{
An object of class \code{character} of length 7.
}
\usage{
whipalette
}
\description{
Use crea_whipala() para construir whipalas de diferentes tamaños
}
\keyword{datasets}
